+++ src/loaders/anndata-loaders/BaseAnnDataLoader.js
@@ -41,7 +41,9 @@ export default class BaseAnnDataLoader extends AbstractLoader {
   decodeTextArray(buffer) {
     return (
       new TextDecoder()
-        .decode(buffer)
+        // Remove header: https://github.com/zarr-developers/numcodecs/blob/2c1aff98e965c3c4747d9881d8b8d4aad91adb3a/numcodecs/vlen.pyx#L34
+        // Should we validate? Seems unnecessary, but maybe?
+        .decode(buffer.slice(4))
         // https://stackoverflow.com/questions/11159118/incorrect-string-value-xef-xbf-xbd-for-column
         // for information on the right hand side of the | in the regex.
         // eslint-disable-next-line no-control-regex
@@ -147,9 +149,7 @@ export default class BaseAnnDataLoader extends AbstractLoader {
             throw err;
           }
         }
-        const text = this.decodeTextArray(data)
-          .filter(i => !Number(i))
-          .filter(i => i.length > 2);
+        const text = this.decodeTextArray(data);
         return text;
       }));
     return this.cellNames;